check_memalign: script to find instances of missing MEMALIGN macros
The idea is to detect all classes or structs that are missing their
MEMALIGN macro, so that we can verify the code is Eigen-safe.

The NOMEMALIGN macro can also be used in select cases where the class is
known not to be affected. Such classes will no show up in the output.
However, this should only be used for sporadic cases, the normal case should be
to declare the class with MEMALIGN.