avoid use of ATOMIC_FLAG_INIT in class initialiser list
This causes clang++ to issue warnings.  According to
http://en.cppreference.com/w/cpp/atomic/ATOMIC_FLAG_INIT,
it seems the only correct way to use ATOMIC_FLAG_INIT is in a construct
of the form:

    std::atomic_flag static_flag = ATOMIC_FLAG_INIT;

"the value held after any other initialization is unspecified."

According to this StackOverflow post:
http://stackoverflow.com/questions/20209021/what-is-the-difference-between-using-atomic-flag-init-and-stdatomic-flagclea,
it's probably just fine to call clear() once constructed. There won't be
any race conditions in our case anyway...