New implementation of .row() method
Also with a commend for testing: cmd/test_row_method.cpp
This version is now safe to use with any ImageType class, regardless of
whether it is accessed with direct IO, and Adapters.

All seems to check out OK. The following types of syntax are possible
(assuming images have size 3 along axis 3):

  out.row(3) = in.row(3);
  out.row(3) += in.row(3);
  out.row(3) -= in.row(3);

  Eigen::Matrix3f M;
  out.row(3) = M * Eigen::Vector3f(in.row(3));
  out.row(3) += M * Eigen::Vector3f(in.row(3)) + Eigen::Vector3f(in2.row(3));

  Eigen::VectorXd x (in.row(3));
  out.row(3) -= x;
  x -= in.row(3);

etc.