Eigen memory alignment: cleaner replacement for std::vector
This replaces my earlier solution for this (memalign_vector) with an
equivalent but syntactically cleaner alternative. I've replaced it with
a new MR::vector<> class, which can be used as a direct replacement for
the std::vector class. It derives from std::vector, using the
Eigen-provided aligned allocator for those types that require it. So it
should come with zero overhead.

I'll add a test to the check_memalign script to catch all occurrences of
std::vector, and basically ban its use outright - that's the only way to
guarantee that the aligned vector is used when required.