Tests: Replace ';' with '&&'
If commands within a test line are separated by semicolons, then only the output status of the final command will be received by the calling script; this is particularly problematic if the final command is e.g. rm in order to clean up temporary files, since the return value of such a command may be 0 even though the actual test failed.