dwi2response tournier: Fix for zero crossings
If running dwi2fod and fod2fixel results in zero crossing fibres detected in all candidate voxels, this change ensures that the algorithm exits appropriately using the result from the previous iteration.
Also ensure that if file error.txt is written, it is not read in binary mode, in order to avoid encoding issues.