MRView: changes to command-line processing
These changes try to force operations to actually happen after each
option is processed, since a lot of things happen during a GL redraw.
Forcing an explicit redraw seems remarkably difficult in Qt for some
reason... This relies on setting up a new timer with a 10ms timeout to
call the slot to process the next option, so that any updates to the
render can also happen before that. Setting a timeout of 0 doesn't work
reliably, sometimes updates don't happen in time for some reason...
There is a therefore a chance that this remains fragile even with a 10ms
timeout, depending on exactly how Qt does its event processing.