Scripts: New function file.waitFor()
The intent of this function is to get a script to halt processing until a particular file not only exists, but a valid file handle can be acquired.
This can therefore be used in cases where e.g. network-based file systems may introduce a race condition within a script, or SGE may by used to execute a particular function but script execution immediately continues as soon as the jobs are sent.
Since this hopefully fixes issues encountered in the 5ttgen fsl script due to the use of SGE by the FSL script run_first_all, SGE has now been re-enabled.