Fixel::find_directions_header() changes
Previously, all images in the fixel directory were opened, and then is_directions_file() was called, which checked both the image properties and the file name.
Apart from being inefficient, this also clogs -info output as the directory is scanned.
This change validates that the name of a file is a plausible directions file before instantiating a Header and checking the image properties.