dwipreproc: Fix bug w. -rpe_pair when PE info present in header
This fixes an issue when the user uses the -rpe_pair option, but phase encoding information is present in the input image headers. The phase encoding table for the topup call was being erroneously overwritten with the contents of the manually-generated table, based on the user's -pe_dir input (but the default total readout time). So the estimated inhomogeneity field would be scaled based on the default total readout time, but the subsequent eddy call would use the phase encoding table from the input DWI header, which contained the actual total readout time. The consequent effect was equivalent to multiplying the inhomogeneity field by a scalar: the EPI inhomogeneity field correction would be in the correct direction, but be either under-corrected or over-corrected.
Various other parts of the script were also cleaned up during the process.