Prevent estimation of optimal threshold when user specifies value
In the original code, the optimal threshold was always being calculated,
even when the user specified the threshold on the command line.  Now,
check that the threshold flag was not given before calculating the
optimal threshold.  Saves some processing time and prevents the user
from being worried that their value was not being honored.