dwipreproc: Additional instructions for -topup_options and -eddy_options
Unfortunately the argparse module appears to stumble if the user provides a quotation-encapsulated string as the value to one of these options, if that string does not contain a space within it. Adding a stray space character, or separating the option string and value with a '=' rather than a ' ', fixes the issue; the former is an easier instruction to give to users, and is more consistent with MRtrix3 command-line usage.