5ttgen fsl: Change waiting for FIRST results
Rather than testing for the presence of every .vtk file output by FIRST, instead look for the combined image created by the FIRST post-processing script. Also provide the user with a more clear warning regarding what it's waiting for, and the fact that there's a chance that FIRST may fail.