fixelcfestats: New option -mask
This option restricts the statistical analysis to a subset of fixels, as defined by a boolean fixel data file.
Calculation of the fixel-fixel connectivity matrix, data smoothing, t-statistic calculation, and stiatistical enhancement, are all only performed on those fixels within the mask.
However, the output fixel directory contains all of the fixels from the input template fixel image, not just those within the mask; this retains fixel-fixel correspondence between the input template image and the output results.
For fixels outside the mask, a value of NaN is written for all fixel data file outputs.