Switch to ConjugateGradient solver.
The purpose of this commit is to save half of the comutational effort involved in cumputing the weights in the design matrix M.

Previously, we used the LeastSquaresConjugateGradient solver to deal with non-square matrices. LCGS is generally beneficial because it avoids calculating M'M. However, to save on memory, we had to compute blocks of the design matrix M online. As such, these blocks were computed twice in every iteration: once to compute M x and once to compute M' y. Since computing these weights is a very expensive operation, it is beneficial to compute them only once per iteration and then provide an efficient, custom implementation of M'M x.