mrtransform & mrresize: Fix nearest-neighbour downsampling
When using nearest-neighbour interpolation and mapping to a lower-resolution image, if the default oversampling strategy is used, this will result in averaging of sub-voxel samples and therefore output intensities that do include some degree of "interpolation" with respect to the input image. This change disables oversampling when using nearest-neighbour interpolation.