Tractography::Properties::get_step_size() function
This function homologates a number of independent pieces of code that try to deal with the fact that the distance between points along a streamline may not be equivalent to the internal step size of the tractography algorithm, nor may such a fixed step size even exist.
It also fixes a bug where running the tckstats command where the step size was not defined (e.g. in the output of tckglobal) would result in an error trying to convert an empty string to a float.