5ttcheck: Relax and clarify warnings / errors
Some users were alarmed by their 5TT images being reported as erroneous, when a single voxel did not have a unity sum of partial volume fractions (which really doesn't matter if it's a single voxel and the error is not too bad). This change relaxes from Exception to WARN() level the message when only a single voxel is erroneous. Furthermore, 5ttcheck will only return non-zero if one or more input images disobeys the fundamental 5TT requirements; if there are only warnings regarding non-unity sums of partial volume fractions, an additional warning will be isued at completion, but the command will still return zero.
In addition, the -masks option behaviour depends on the usage: If a single input image only, the raw string will be used as the output image name if it constitutes a valid image name, otherwise ".mif" will be appended; only if there are multiple images will the name of the corresponding input file also be appended to the output image path.