Testing: Use pylint
New script run_pylint in root directory runs pylint on all Python files in bin/ and lib/.
This is included in the Travis CI tests, and should be tested independently in Python versions 2 and 3.
File testing/pylint.rc specifies what types of warnings / errors will be reported by pylint and hence result in testing failure.
There are also a wide range of fixes that have been applied to the repository in order to pass the majority of tests; remaining issues are mostly to do with pylint usage (reporting library functions as unused variables), and accessing protected members from mrtrix3.app, which should be refactored.
Discussion ongoing in #1074.