docs: move ARCH discussion out of standalone packager
This section about setting the CPU architecture was included in the
standalone installer instructions due to the fact that we used to run
with -march=native - this was likely to cause issues for this type of
installation specifically. Now that the default is up to the compiler,
there is no longer any reason for this section to be where it is.