Changes to enforce requested thread counts
At the track mapping stage of tcksift / tcksift2, multiple threads were being invoked for both the streamline-voxel mapping and the streamline-fixel visitation encoding, resulting in twice as many threads being spawned as what was requested. These two tasks are now combined into a single functor such that the requested number of threads is honoured.
In tckgen -seed_dynamic, multi-threading both the tracking and streamline-voxel mapping functors could result in the use of more resources than had been requested. In these changes, the back-end keeps track of whether the number of threads was explicitly requested by the user, or whether it was set automatically based on the detected hardware concurrency: If an explicit request was made, then each of these stages will run with half of the requested number of threads, such that the total requested resources are not (substantially) exceeded; whereas if no explicit request is made, then the maximum number of threads will be spawned for both stages, with the system then being responsible for managing the load between the different threads.
Relates to discussion in #1105.