Scripts: Improved Ctrl-C handling
If a keyboard interrupt occurs during standard processing, it should be caught by the custom signal handler, appropriately shutting down the script. If however it occurs during a run.command() call, it is necessary to instead catch the KeyboardInterrupt exception, such that any I/O calls go out of context and the subprocess threads can be properly shut down.