fsl.findImage(): Fix for MRtrix3_connectome script
When the image being sought resides in a directory that contains a period character (such as the default produced by the fsl_anat script), the original function would extract the basename of the directory path, rather than the basename of the image file. This change ensures that the full directory path to the image being sought is included in the search path.