split(): Ignore leading delimiters
When the "ignore_empty_fields" parameter is set to true, one would expect to not have any empty strings present in the output vector. However, if the input string commenced with a delimiter character, an empty string would be written as the first element in the vector. This change ensures that when "ignore_empty_fields" is set to true, any leading delimiters are ignored.