dwigradcheck: Fix dwi2maskcall
In use case where input image header does not contain a gradient table, and user instead relies on the -grad or -fslgrad option, dwi2mask would fail due to the absence of a gradient table.
Instead have dwi2mask explicitly import from grad.b, which is guaranteed to exist regardless of how the baseline gradient table was provided.
No need to import the gradient table into the image heder for the rest of the script, since tckgen is always run with an explicit -grad or -fslgrad option based on a permuted gradient table.
Supersedes #1151.