dwipreproc: Support eddy slice-to-volume correction
If the user specifies --mporder within -eddy_options, this will be handled in one of two ways:
- If the user has also specified --slspec within -eddy_options, then dwipreproc will find that file, copy it into the temporary directory, and modify the contents of -eddy_options to reflect the location of that file.
If --slspec is not specified within -eddy_options, then dwipreproc will look for 'SliceTiming' within the input DWI header key-value entries. If this is found, then an appropriate slspec file is generated based on the slice timings, and this is provided to eddy.
Closes #1169.