More playing with TravisCI build matrix
By using an if statement within .travis.yml, return codes of test scripts would normally be ignored. A proposed solution to this was to trail each test with " || travis_terminate 1". Unfortunately this applies a hard termination, and the "after_failure" section of .travis.yml (containing the details of any errors) could not be accessed.
This change moves the test branching into a dedicated script, travis.sh, which is configured to error out whenever any command executed returns an error code.