tcksift2: New option -linear
The code for this feature was added a long time ago, but command-line access to the functionality was never provided.
When using the -linear option, the command estimates a weighting factor for each streamline in a single step, with each factor determined independently. This is a considerably different approach to the primary SIFT2 optimisation algorithm, but nevertheless operates on a similar principle, and with the same fundamental goal. Its main use is in the conceptual demonstration of why the SIFT2 algorithm mis designed the way it is.