dwipreproc: Compatibity for --mporder and padding
If the number of slices in an axis of the image is odd, and topup is being used, then dwipreproc will duplicate the final slice along that axis, in order to prevent a fatal crash in topup.
However, if slice-to-volume correction is being used, and there are an odd number of slices, then this padding will result in a mismatch between the number of slices in the image data and the number of entries in the slice timing data.
These changes compensate for this by correspondingly altering the slice timing information, and delaying writing that information to file until after image padding has taken place.