tckgen: Recursive back-tracking
First attempt at making the ACT back-tracking mechanism recursive.
Whenever a track is truncated and tracking re-attempted, those attempts are themselves permitted to make use of back-tracking.
Initial evidence suggests that this mechanism provides no tangible benefit, and possibly runs slightly slower / requires more restarts, compared to the existing approach, if only 1 attempt per truncation length is made.
If the recursive algorithm is permitted to make 3 attempts per truncation length, tracking may freeze; it is not yet known whether this is due to a bug in the implementation, or just the exponential increase in recursion complexity.
Changes are being committed in case this concept is revisited at a later date.