Stats: First compilable version with Freedman-Lane
- Modify how multiple contrasts are handled. Initial support for this was previously implemented for vectorstats only, based on the contrast being a matrix rather than a vector. However, this framework does not naturally extend to F-tests. This new approach stores each within a GLM::Contrast class, and these are explicitly looped over whenever required.
GLM t-test code has been replaced with the F-test (i.e. without variance groups) as presented in Winkler et al., 2014. This has removed many optimisations, but was necessary in order to sufficiently generalise the framework for upcoming enhancement.
Changed stdev calculations in statistical inference commands to a vector, since this does not change betwee contrasts.
- Removed apparent redundant creation of output images related to the default permutation in mrclusterstats.
- Model is currently partitioned based on null columns of the contrast matrix only.
Note that this is merely the first COMPILING version of this code.