FOD FMLS: Apply peak amp threshold after peak revision
This means that if the maximal amplitude value sampled for a lobe from the dixel set is just below the threshold, but Newton revision on the sphere yields an amplitude that is just above the thresold, the fixel survives the threshold test.
Also removed multiplie redundant calls to FOD_lobe::finalise().
Identified from discussion in #1228.