Tracking code: Proper use of polymorphism
Previously, some tracking code was heavily reliant on the use of templating in cmd/tckgen.cpp. While the code would behave as expected in this context, some aspects appeared as though they had been designed with polymorphism in mind, but would not exhibit the expected behaviour if actually used in this fashion. This change introduces the correct polymorphism concepts such that the code can be used in this way if desired. It however does not have any significant influence on execution speed of tckgen, since it still exploits the template parameter of the Tracking::Exec class and hence does not use the underlying polymorphism.
Addresses #1234.