Permit writing of empty tracks to file
Rather than the (long-standing) behaviour of interpreting empty streamline data as a 'failed' streamline that only contributes to the 'total_count' header field, this change requires that an explicit skip() function be called to increment the 'total_count' field when appropriate, and any call to the standard Tractography::Writer functor will be written to file, even if the number of points in the streamline is zero.
Currently untested.
As discussed in #1131 and #1154.