configure: Alter debug optimisation on Windows
On Windows, when performing a debug build (i.e. '-g -O0'), compilation fails in multiple locations due to file size limitations.
This partially reverts the change of optimisation level when compiling in debug mode from commit fa997ca, discussed in #1072 and merged in #1143, only reverting back to -O1 for Windows.