FBA docs update: actively recommend to not use the -fsl option to dwibiascorrect
This option is not compatible with the intended pipeline, where the mask will be reestimated later.  I know a user could have gotten masks from another source beforehand, and supply them to the script, etc... but such a customisation of the pipeline goes beyond the basic instructions as provided.  These instructions should be as clear as possible to avoid issues by all means.