remove template warp shortcut from FBA pipeline
This was introducing more issues for users than it was worth. I've just witnessed a collaborator accidentally mixing this with the separate registration step; i.e., building a template out of a subset of all subjects, using those warps, and for the remaining subjects doing the registration to get their warps. However, the regularisation and qualities of the latter warps ends up being different to the former, hence introducing a bias in the study. Also had a collaborator being surprised that doing the shortcut vs not doing it, resulting in slightly different outcomes (and effect sizes for FC). The step is not worth the potential for these confusions and ultimately mistakes, since the time the explicit registrations take in the FBA pipeline is easily an order of magnitude smaller than the time it takes to build the FOD template (since that's an order of magnnitude repeated registrations for all subjects that could otherwise benefit of the shortcut).  If people feel confident, they can still tweak the pipeline, but I'd rather have the docs be as robust as they can be to avoid mistakes for less experienced users.