tckresample -num_points: Fix rare error
In cases where a particular streamline vertex is duplicated (as appears to be the case in the testing data), the distance between successive vertices will be zero; this could lead to errors in the streamline resampling process, which relies on calculating interpolation coefficients between vertices based on their relative distances. This was caught due to the assertion checks for finite track & track scalar data added in 2d01ba6f.