Scripts: Fix erroneous error report
If a script is re-run using the -continue option, and the -nocleanup option is also used, then the file 'error.txt' resulting from the prior script failure will be detected and erroneously echoed to the command-line as an error. This is solved simply by erasing file 'error.txt' when the -continue option is used.