Command-line usage: Warn on likely erroneous option usage
If one of the inputs to a command-line option starts with a dash character, it is likely that the contents of that command-line entry are intended as its own command-line option, and are not supposed to be an input argument to some other command-line option. If this occurs, the most likely scenario is that the user has not provided enough arguments to the first command-line option, and hence the second command-line option is being 'consumed' as one of the requisite inputs to the first option.