tckmap: Fix segfault with -dec -stat_vox mean combination
Fixes issue introduced in 17e22166. The specific use case of combining -dec with -stat_vox mean was altered to no longer rely on keeping track of a voxel-wise track density, and so memory for this buffer is no longer allocated; yet the receiving functor would still attempt to write to this buffer.
Closes #1285.