mrview Overlay tool: Always clear & select on overlay load
Whenever one or more overlay images are loaded, the current selection within the overlay image list will consist only of the just-opened images.
In the case of use of the -overlay.load option, it is always only the most recent instance of the -overlay.load option to which subsequent -overlay.* command-line options will apply.
As discussed in #1170.