Function MR::check_app_exit_code()
The goal of this function is to attempt to detect instances where File::MMap::~MMap() has failed, but written a non-zero value to MR::App::exit_error_code rather than throwing an exception. While modifying this variable will result in the command yielding a non-zero return code, other parts of the command may nevertheless continue to execute despite such an error having occurred. This function may permit prevention of further execution in cases where such an error has occurred; on the condition that this function (or anything that invokes it) should never be called from a class destructor. Relates to discussion in #984.