Prevent erroneous command-line usage warnings
In #1282, a command-line warning was introduced when the value provided for an argument associated with a command-line option commenced with a '-' character, as this is likely an indication that the user has not provided enough input arguments to that option and hence a later command-line option is being erroneously assigned as an argument to the prior option. However this fails to take into account the possibility for negative numbers to be provided as such arguments. This change therefore suppresses this warning if a particular option argument is flagged as any numerical type.