dwipreproc: Remove -cuda option & capture topup/eddy outputs
The script (and associated library function) has been modified in order to automatically attempt to use the cuda version of eddy, without requiring the user to explicitly specify the -cuda command-line option. If the cuda version is not run sucessfully, or is not available, then the OpenMP version will then be executed.
Inspired by discussion in forum thread: http://community.mrtrix.org/t/dwipreproc-eddy-cuda-slspec-unable-to-open-file-multibandgroups/1573/7
In addition, capture the output text generated by the topup and eddy commands, and write them to text files within the temporary directory; while these will typically be deleted along with the temporary directory, it may be useful for some users who wish to access this information by using the -nocleanup option. This was also suggested on the forum: http://community.mrtrix.org/t/capture-standard-output-of-topup-and-eddy-in-dwipreproc/1284