New argument type type_various()
Previously, the argument tyep type_text() has been used as a fall-back position when the input to an argument may be one of various types (e.g. either a path to an image or a comma-separated list of values). However this introduces limitations in the context of #1282 and subsequently #1293, where argument values are tested to see if the first character is a '-', which produces a warning for the user regarding potential mis-specification of command-line inputs. #1293 involves fixing the fact that any numerical input may begin with a dash character, and hence must be excluded from this test. By providing a type_various() that is separate from type_text() (which should now be reserved purely for string inputs e.g. output filepath prefixes), arguments flagged as 'various' may potentially contain a numeric input and are hence excluded from this test, whereas arguments flagged as 'text' will still be tested (since even if the input to such an argument can be arbitrary text, choosing to commence that input with a dash character is almost certainly a bad choice, and therefore a warning is probably warranted).