run.command(): Consistent stderr return decoding
When running with -debug option, use the Windows1252 codec to decode byte data for text display on the terminal, but additionally decode the entire stream contents using UTF-8 after the command has completed. This ensures that the returned stderr contents within a script will not vary between use or non-use of the -debug option.