mrview Connectome tool: Change to edge-visibility-by-nodes handling
When attempting to build a demonstrative example for a course, I discovered that the capability to influence the visibility of edges based on the visibility of nodes was lacking. The "Edge visibility" combobox had an option to toggle edge visibility according to which nodes were visible (only edges where both connected nodes are visible are also visible), but this then prevented e.g. setting a threshold on edge visibility based on connectome edge strength.
This change removes that option from the edge visibility combobox, and instead introduces an additional checkbox, which determines whether or not the requirement of having both nodes associated with an edge to be visible in order for the edge to be visible should be applied.