tckedit: Fix output
The class responsible for receiving streamlines data processed as part of the tckedit command was not appropriately updated to reflect the new capability of the Tractography::Writer class to write streamlines with no vertices in #1259, specifically commit fc40ee8. This change ensures that zero-length streamlines are no longer erroneously written to the output file, and the 'count' and 'total_count' fields in the track file header are calculated appropriately.