mrview ligthbox mode: extensive overhaul
This gets rid of a lot of cached data (projections, mostly), in favour
of on-the-fly recomputation. There are also changes to the way the
'step through volumes' feature is handled, notably to ensure that the
current volume (as shown in the view tool) is the volume under the
crosshairs.

Still to do: windowing reset doesn't work as expected, presumably due to
min/max values not being cached per volume. I'll leave this for some
other time...