Prevent erroneous command-line usage warning
In #1282, a terminal warning was introduced in an attempt to catch cases where users failed to provide enough input arguments to command-line options.
This was refined in #1293 to prevent erroneous warnings from being displayed when having an input argument begin with a dash character is entirely reasonable.
This change further refines the logic dictating when a terminal warning should be issued, by identifying the use of piped images with command-line option inputs and not issuing the warning in that circumstance.