GLM: Addtional restrictions on per-element design matrices
When the number of subjects contributing to the statistical test within an element is greatly reduced due to non-finite values, it becomes possible by chance for the model to obtain an exceptionally good fit to the data; this leads to an exceptionally high t-value / F-value, which correspondingly leads to very slow execution during statistical enhancement. This change introduces an additional constraint thta must be satisfied in order for statistical testing to proceed for any given element: If at least one subject is excluded in that element, the number of remaining subjects must be at least twice the number of factors inthe design matrix, in order to provide adequate degrees of freedom.