dwi2response: Fix for absent -mask and no gradient table
If a mask is not provided to the script via the -mask option, the script will automatically generate a brain mask using the dwi2mask command. However, this command was being invoked upon the input image as provided by the user. If this image did not contain a gradient table in the header, and the user was relying on the -grad or -fslgrad option to provide this information, the dwi2mask call would fail. This change ensures that the dwi2mask command is executed on the DWI that has been imported into the temporary directory, and therefore is guaranteed to contain a diffusion gradient table.