mrview Connectome tool: Fix edge visibility
In b2675375, controlling of connectome edge visibility was altered due to an observed inadequacy when attempting to suppress the display of edges based on the visibility of the nodes to which they are connected. However, removal of the relevant group box entry for some reason was not included in the commit, leading to the behaviour of each entry in the group box being inconsistent with its description. This change rectifies the mismatch.