Scripts: Tweak message on command failure
Users seem to continue reporting failure of commands executed within a wrapping script as though they are failures of the script itself. Maybe a minor rewording of how the error message is presented will reinforce the fact that in many circumstances, such commands are responsible for their own failure, rather than the script providing the command with bad data resulting in a failure.