mesh2voxel: Tweak to partial volume estimation
In some cases, a surface may intersect a single voxel at two locations that are a long distance from one another on the surface. When this happens, the directions of the normals of those polygons may be almost directly opposing. In this case, one does not want to choose whether or not a particular point is inside or outside the surface based on the polygon to which the point most accurately projects to the inside of the polygon; instead, if the point projects to the inside of more than one polygon, the polygon to which the point is geometrically closest should be used to determine whether the point is inside or outside the surface.