Surface::Filter::Smooth: Skip tiny surfaces
Attempting to smooth a very small surface (e.g. 1 voxel from voxel2mesh) seems to sometimes result in erroneous behaviour in a subsequent mesh2voxel call, where the entire image is labelled as being inside the surface. This is a quick fix to prevent this behaviour where it's been observed, but if a similar result is observed elsewhere then more comprehensive digging will need to be performed.